Create UserList component
