+++ components/index.js
@@ -1,7 +1,9 @@
 import Confirmation from './confirmation'
 import ProductList from './product-list'
+import UserList from './user-list'
 
 export {
   Confirmation,
   ProductList,
+  UserList,
 }
