+++ components/user-list/index.js
@@ -1,0 +1,3 @@
+import UserList from './UserList'
+
+export default UserList
